F-3 Refactor
