+++ src/App.css
@@ -1,4 +1,4 @@
-.App {
+.tictactoe {
   text-align: center;
   display: flex;
   flex-direction: column;
